// apps/react-frontend/src/api/client/models/truck.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Truck
 */
export interface Truck {
    /**
     * 
     * @type {string}
     * @memberof Truck
     */
    'licensePlate': string;
    /**
     * 
     * @type {string}
     * @memberof Truck
     */
    'brand'?: string;
    /**
     * 
     * @type {string}
     * @memberof Truck
     */
    'model'?: string;
    /**
     * 
     * @type {string}
     * @memberof Truck
     */
    'updatedAt': string;
    /**
     * 
     * @type {string}
     * @memberof Truck
     */
    'displayName': string;
}